### 2020-07-19 #2429 @richardjgowers
@valeriewelborn how modified is the patched version of TINKER (https://github.com/WelbornGroup/Tinker_ELECTRIC) from the original?  Also, I'm not sure but is the rehosting of modified source code allowed according to the license? (in particular points 5&6 of https://dasher.wustl.edu/tinker/downloads/license.pdf)
