### 2020-07-20 #2429 @janash
Hi @richardjgowers - The Tinker software has switched in recent years to be distributed via GitHub (https://github.com/TinkerTools). The modified Tinker code included here is part of a larger project at The Molecular Sciences Software Institute called the MolSSI Driver Interface (MDI).  You can see the MDI (which has the same code as Tinker-ELECTRIC) here - https://github.com/MolSSI-MDI/Tinker/tree/mdi-ef . This repo has a commit history compatible with the main Tinker repository. We can change the link in the README for this repository to this fork if that is helpful.

We have had discussions with the Tinker core developers about the Tinker-MDI project and plan to eventually have this code incorporated into the main Tinker repo. We can contact the Tinker developers to ok this repo if necessary.

As far as changes to Tinker, the main changes are the addition of the MDI library (in folder `mdi` in Tinker-ELECTRIC). We also have modified some of the Tinker source code. We have added the file `mdiserv.f` to enable movement of information out of Tinker. The Tinker software calculates the electric field for each atom in order to calculate dipoles for atoms. This is usually internal and not saved. We have added the file `efield.f` and modified `induce.f` in order to capture this information. Currently, Tinker is the only software which calculates the electric field in this way. However, the software submitted here (the ELECTRIC driver), may be used eventually with other codes that can output electric field information.

Tagging @taylor-a-barnes in case I've forgotten anything.
