### 2020-08-03 #2429 @danielskatz
@janash - sorry for the delay - after some editor discussion, we have decided that this submission is not a large enough contribution for JOSS to review it - see https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort. We encourage you to consider JOSS at a later time if more effort and code goes into this software, or alternatively, consider publishing the software to Zenodo to get a DOI that users can cite.
