### 2020-08-03 #2429 @janash
Hi @danielskatz and @richardjgowers  - Thank you for your consideration on this paper.  In your requirements for publication, I see that you have guidelines about code length. I understand that in general a smaller number of lines of code may represent less effort. However, for this project, a small(er) number of lines indicates careful consideration and design (or effort).

I disagree that the number of lines of code for a project is reflective of the amount of scholarly effort. This project could have been much larger (more lines of code) if it did not use an efficient and systematic approach to the problem. We implemented this is a clean and compact way, and in my opinion that should not negatively reflect on the software.

This is a useful analysis tool for researchers working in molecular simulation, and enables an analysis ability that is **not possible in any other existing open source software**. The closed source version of the code is not openly available, actively maintained, updated, or tested.

This project, despite its appearance of small size, represents a significant effort and is not something that could not be easily replicated. This project required the use of multiple tools, programming languages, and knowledge of scientific concepts. I am outlining notable points below, though there are several other concepts and tools at play:

1. Understanding the structure of the Tinker software package (written in Fortran). We had to decide what information was relevant, where it was stored, and how to retrieve it. The Tinker source code was modified so that information that was previously discarded was recorded and made available in a useful way.
1. Modifying the Tinker software package to use the MolSSI Driver Interface (MDI) which allowed efficient communication of data and structures inside of the Tinker software to be available.
1. Efficient analysis of retrieved data using `pandas` and parallelization of the analysis using `MPI4py`.

This tool adds the ability to customize that analysis by atom, molecule (as counted by Tinker) or residue (as specified in a pdb file) without the user editing code.  Our implementation is robust, public, and much faster than the previous closed package which was used for analysis. Our code is also parallelized with very high efficiency, so the achievable wall times are dramatically lower than what you could get from any serial code. Our benchmarks of our parallelized code showed an order of magnitude improvement over the serial version.

In summary, we believe that this project represents significant effort appropriate for JOSS. Again, thank you for your consideration, and we hope for productive future collaboration.
