### 2020-08-04 #2429 @valeriewelborn
Hi all,

I would like to second Jessica's comments and I would also argue that it follows the rule of thumb "no less than 3 months of work for an individual" (which translates to different real time depending on the skills of the coders - a good graduate student would have taken at least 6 months to produce a working code but likely not of the same quality). I also want to emphasize, as Jessica mentioned, that the calculation of electric fields from Tinker Amoeba calculations is not trivial and currently not available to the community. This would be the first time such a tool would be made open source. Using electric fields to analyze MD trajectories would benefit the community tremendously as I have demonstrated in Nature Catalysis (https://doi.org/10.1038/s41929-018-0109-2), JACS (https://doi.org/10.1021/jacs.9b05323) and ACS Catalysis (https://doi.org/10.1021/acscatal.7b03151). I also explained the underlying principles in a Chemical Reviews (https://doi.org/10.1021/acs.chemrev.8b00399).
