### 2020-08-04 #2429 @danielskatz
Let's make sure we are talking about the same thing.  In this repository, I see
 - https://github.com/WelbornGroup/ELECTRIC/tree/master/ELECTRIC (a new contribution here)
 - https://github.com/WelbornGroup/ELECTRIC/tree/master/ELECTRIC/mdi (which seems to be a copy of https://github.com/MolSSI-MDI/MDI_Library)

What did you intend to submit for review?
