### 2020-08-06 #2429 @janash
The ELECTRIC repository is what we intended to submit for review.

https://github.com/WelbornGroup/ELECTRIC/tree/master/ELECTRIC/mdi is a copy of https://github.com/MolSSI-MDI/MDI_Library. We do not intend for this to be considered in the review.

The parts of the repository specific to the ELECTRIC driver in this directory (https://github.com/WelbornGroup/ELECTRIC/tree/master/ELECTRIC), excluding MDI.

This analysis requires use of MDI-enabled Tinker (Tinker-ELECTRIC), linked in the submitted repository. You can see our version with commit history here - https://github.com/MolSSI-MDI/Tinker/tree/mdi-ef . This modification to Tinker was a large part of the work for this project. Once we were able to get the information we needed via this MDI enabled version of Tinker, the driver itself (in the ELECTRIC repository) was not too much trouble to write. We did work to make the implementation more convenient (via use of` argparse`, and ability to specify molecule, residue, or atom analysis) and efficient (parallelization using `MPI4Py`).

Concerning mentioning 'design' and LOC in my earlier comment - When we started this project, we knew in general what we wanted to calculate (the electric field), and that electric fields are calculated (in some cases) in Tinker in order to calculate multipoles. What was less clear was at what stage we retrieve information from Tinker or what calculations we should do ourselves. One option would have been to get multipole information and coordinates and do the mathematically intensive calculation outside of the software in our own code. This would have resulted in a much more complicated piece of software (probably less efficient, more prone to error, more lines of code, and yes, much more work). However, there is an existing subroutine in Tinker that does calculate the electric field as we needed it to be calculated, but it is not visited in a typical molecular dynamics run. We decided to send coordinates from a trajectory to Tinker using MDI and to force the calculation to go through this subroutine. The information we needed to retrieve was still not readily available to be passed out of Tinker using MDI as it is discarded (the electric field is typically only used as a means to calculate multipoles and not used for analysis, and is not calculated on a per atom basis under typical conditions). So, we also had to alter the Tinker source code to visit the relevant subroutine when we needed, retrieve, and pass out this information.

At this point we also had to choice to either make this into something that ran alongside the MD run or a post processing method. We chose to make this a post processing methods so that we could parallelize this extra calculation.

Once we could retrieve the electric field, it was a matter of summing the information correctly in the driver (so I understand how `ELECTRIC.py` doesn't look like much).  Considering the size of trajectories and amount of information we retrieve from Tinker, we also made some choices in `ELECTRIC.py`. We read in trajectories in chunks by frame since some trajectories can be quite large. Since this is a post processing method and the trajectory frames are already generated, calculation of the electric field for each frame is independent (making it 'embarrassingly parallel'), so we have used MPI4Py to parallelize. This is, of course, much faster than doing the calculation in tandem with the simulation. The use of `argparse` was also strategic as these type of calculations will typically be submitted to a queuing system in a script similar to `tcp.sh` (included in the repository).

Just wanted to explain some of the design choices and considerations we went through during implementation that are not reflected in the LOC of the final product. None of the choices we made were clear initially. What we ended up with is a relatively small piece of code which runs quickly and delivers novel analysis for researchers in this field. As @valeriewelborn mentioned, this type of analysis has already been shown to be useful in a number of peer reviewed publications, and was previously not available to most researchers. We felt this was notable and worthy of a JOSS submission. Thanks again for your time and consideration.
