### 2020-08-06 #2429 @taylor-a-barnes
@danielskatz Thank you very much for the time you’ve devoted to clarifying the nature of this submission. As the above discussion suggests, the boundaries defining exactly what constitutes this submission are not entirely trivial. To be clear, any calculation involving the ELECTRIC code involves participation of the following components:

1. The ELECTRIC code, which we originally submitted for review.
2. A version of the Tinker code which was modified as part of the ELECTRIC project (https://github.com/MolSSI-MDI/Tinker/tree/mdi-ef); we intend to create a PR for inclusion of these changes into the official Tinker repository. Out of the (perhaps misguided) expectation that including those modifications in the review would complicate the process, we did not originally intend for these contributions to be included in the review; however, if doing so would be useful for clarifying the scope of the project, we would be willing to have the Tinker contributions reviewed as well.
3. The MDI Library (https://github.com/MolSSI-MDI/MDI_Library), which we use to handle efficient communication between ELECTRIC and Tinker. An internal distribution of the MDI Library is included in the ELECTRIC code, but it should not be considered as part of this submission.

I can understand that the submission, as originally presented, might not appear to fulfill the JOSS requirements for “substantial scholarly effort” (although, as @janash points out, the scholarly effort associated with the ELECTRIC code is much larger than the LOC count might imply). Nonetheless, I believe that if the ELECTRIC code and the Tinker contributions are considered in combination, we unambiguously meet the JOSS standard. Our contributions to Tinker easily exceed 1000 lines (https://github.com/MolSSI-MDI/Tinker/blob/mdi-ef/source/mdiserv.f alone is 921 lines), and both @janash and I have made meaningful contributions to the project since January, 2020. Neither the ELECTRIC code nor the Tinker contributions could be considered “one-off” modifications; in particular, the Tinker contributions have been made with the intention of supporting general communication with external codes and are not limited to supporting ELECTRIC.

The primary questions I have are as follows:
1. Is it possible for our Tinker contributions to be considered in addition to ELECTRIC in a JOSS review? If so, are there any additional requirements we should be aware of? For example, would the Tinker contributions need to be presented in the form of a PR?
2. Would you consider a submission that included both ELECTRIC and our Tinker contributions to be sufficient to meet the “substantial scholarly effort” requirement?
