### 2020-08-06 #2429 @danielskatz
You have convinced me (and other editors involved in the discussion).  I've reopened this and am turning it back over to @richardjgowers, who had previously agreed to edit it.

And specifically:

> Is it possible for our Tinker contributions to be considered in addition to ELECTRIC in a JOSS review?

yes

> If so, are there any additional requirements we should be aware of? For example, would the Tinker contributions need to be presented in the form of a PR?

The reviewers will have to identify and understand what those contributions are. A PR would be one way to do this, but this is up to you, I think.

> Would you consider a submission that included both ELECTRIC and our Tinker contributions to be sufficient to meet the “substantial scholarly effort” requirement?

yes
