### 2020-08-15 #2429 @amandadumi
Hello @richardjgowers. Yes, this looks very interesting and well suited for my background!
