### 2020-08-15 #2576 @whedon
[REVIEW]: ELECTRIC: Electric fields Leveraged from multipoleExpansion Calculations in Tinker Rapid Interface Code
**Submitting author:** @valeriewelborn (<a href="http://orcid.org/0000-0003-0834-4441">Valerie Vaissier</a>)
**Repository:** <a href="https://github.com/WelbornGroup/ELECTRIC" target ="_blank">https://github.com/WelbornGroup/ELECTRIC</a>
**Version:** v1.0.0
