### 2020-08-15 #2576 @richardjgowers
Hi @amandadumi and @govarguz thanks for agreeing to review, we'll be doing that in this issue.  In general it works best if you raise issues against the ELECTRIC repository when you find issues.  Let me know if you run into any issues.
