### 2020-08-27 #2576 @govarguz
Hi @whedon, the invitation link has expired. I think, I have accepted it but I changed recently of terminal. Do I need to accept the review invitation again?
Thanks for your attention, Horacio.
