### 2020-09-11 #2576 @amandadumi
Hello @valeriewelborn, thanks for your patience. I am nearly done! Just writing up some final comments which should be posted by tomorrow. I appreciate your patience.
