### 2020-09-12 WelbornGroup/ELECTRIC#17 @amandadumi
JOSS: Documentation Review
This is part of a review for openjournals/joss-reviews/issues/2576

**From JOSS:** Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
Minor corrections:
- The installation instructions in the README.md have the word numpy repeated in a sentence (line 19).
- I feel as though describing the complete relative path throughout the installation documentation is more clear, i.e. "Similarly, edit the file `[...]/ELECTRIC/ELECTRIC/test/locations/Tinker_ELECTRIC'"

Questions:
  - The instructions provided by the authors are clear so this **is not** necessary for approval on my end. Is there a reason why MDI & Tinker_ELECTRIC are not set up utilizing submodules for this project? There could easily be something I'm overlooking. However, if this would work it would be a clean way to keep all parts up to date. Additionally, submodules would allow for an easier setup. I played around a bit to see if this would work on [this branch](https://github.com/amandadumi/ELECTRIC/tree/submodule_build), with an example [build script](https://github.com/amandadumi/ELECTRIC/blob/submodule_build/build.sh).

**From JOSS:** Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)
- The `utils.py` are well documented, Indeed a bit more time to spend with `ELECTRIC.py` before checking this box.

**From JOSS:** Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
- No clear guidelines are listed at present . One suggestion would be to build of the contributing page based off of the [MolSSI cookie cutter](https://github.com/MolSSI/cookiecutter-cms/blob/master/%7B%7Bcookiecutter.repo_name%7D%7D/.github/CONTRIBUTING.md).
