### 2020-09-12 WelbornGroup/ELECTRIC#18 @amandadumi
JOSS: Software Paper Review
This is part of a review for openjournals/joss-reviews/issues/2576

The paper was well written and provided  clear background that was appropriate for non-experts, but the references were appropriate to allow the reader to gain a further understanding of the field and the related software.
Minor change:
 At the end of the summary, the last sentence reads "ELECTRIC was designed to improve quantitative system...", though this seems more of an extension of Tinker capabilities and not  an improvement of accuracy to any existing analysis within Tinker. Thus instead using a wording of "ELECTRIC was designed to expand quantitative system...".
