### 2020-09-12 #2576 @amandadumi
Great work! This is a nice addition to the analysis tools for force fields, especially one that is open source and handles large data efficiently. I have submitted some issues for feedback according to specific section of the review WelbornGroup/ELECTRIC#17, WelbornGroup/ELECTRIC#18.

I have one general question about the Tinker modified for ELECTRIC. The ELECTRIC repository references this adapted Tinker: https://github.com/WelbornGroup/Tinker_ELECTRIC while in the pre-review (openjournals/joss-reviews/issues/2429) the authors mention https://github.com/MolSSI-MDI/Tinker/tree/mdi-efr as the adapted Tinker code. Which of these is correct? If it is the WelbornGroup version, since it is not a fork from the original Tinker repository, it is hard to tell what has changed. Could your provide some insight on where your contributions are? Once I understand which modified Tinker to use, I will provide feedback on the code.

In regards to the  Contribution and authorship point within the General Check category, the submitting author, @valeriewelborn  is not the main code contributor of ELECTRIC according to the git commit log. However, it seems to be that they are the PI leading this work so I do not see an issue with this in terms of publishing as this is very common dynamic. Though with the specific wording of this point, I wanted to ensure this will not bring about any issues for JOSS that I may be overlooking.  Is this acceptable for JOSS, @richardjgowers ?
