### 2020-09-14 #2576 @valeriewelborn
Hi @amandadumi, thanks for the comments.  https://github.com/WelbornGroup/Tinker_ELECTRIC is a copy of https://github.com/MolSSI-MDI/Tinker/tree/mdi-efr so you can go directly to https://github.com/MolSSI-MDI/Tinker/tree/mdi-efr for review of the modified version of Tinker.
