### 2020-09-14 #2576 @govarguz
Dear all, I am looking now to more details but may need a couple of days to
wrap up my comments. Best!

On Mon, Sep 14, 2020, 23:13 Valerie Welborn <notifications@github.com>
wrote:

> @govarguz <https://github.com/govarguz>, @richardjgowers
> <https://github.com/richardjgowers> have you had a chance to look at the
> submission?
>
