### 2020-09-16 WelbornGroup/ELECTRIC#19 @govarguz
JOSS: ELECTRIC review
This is part of the review process for [openjournals/joss-reviews/issues/2576](https://github.com/openjournals/joss-reviews/issues/2576)

The Ms. describes the functionalities of ELECTRIC in a very succinct manner, which also shows the strong background of the Authors in the field.
I recommend addressing the following points before publication.

1. In the last lines of the SUMMARY a phrase reads "_ELECTRIC enables splitting of the total field into contributions from different components of the system, by molecules or by residues as specified in a reference PDB file_". However, looking at the tests or examples @ELECTRIC_github, I was not able to see any example using a reference PDB file. In fact, since PDB is one of the most used formats in biomolecular simulations an example using this format would fit nicely to the description and the whole submission.

2. Within the summary, and as understood that ELECTRIC has a couple of dependencies,  I would suggest to include a flowchart diagram. This may facilitate also the awareness of ELECTRIC to any potential users.

Minor
3. I would also recommend to have a Jupyter Notebook example, powered e.g. by myBinder to show the clear capabilities of ELECTRIC.

JOSS related points, I want to remark and point-out on the Community guidelines and also an outlook of the project. In order to bring the attention of the community and perhaps join projects that may overlap, as we know the molecular simulation community grows fast.
