### 2020-09-16 #2576 @govarguz
@valeriewelborn I have submitted my review to your JOSS Ms. I have made 3 suggestions to the Ms. as seen in [here](https://github.com/WelbornGroup/ELECTRIC/issues/19). Those points may also clarify the concept of ELECTRIC and what is behind. Moreover, clearly show where the ELECTRIC contribution is (e.g. through a flowchart diagram).
Another very minor comment for the Authors affiliation, do you normally use the name without an address? cc: @richardjgowers
