### 2020-09-16 WelbornGroup/ELECTRIC#17 @janash
Hi @amandadumi - thank you for the suggestion of using submodules. I think this simplifies things nicely and is a good way to do installation.
