### 2020-09-19 #2576 @richardjgowers
@valeriewelborn did you mean https://github.com/MolSSI-MDI/Tinker/commits/mdi-ef ?  In general it's useful to see the commits that are "on top" of the base TINKER.  E.g. if https://github.com/WelbornGroup/Tinker_ELECTRIC could have one commit which is the check in of the release of TINKER you built on and another commit (or more) showing the changes made for this work.  That way someone can quickly review how/why the dynamics are different (or are the same).
