### 2020-09-19 #2576 @richardjgowers
@valeriewelborn also I think we'll need full addresses in the affiliations if possible, thanks
