### 2020-09-24 WelbornGroup/ELECTRIC#19 @janash
Hi @govarguz - I am working on these suggestions. Can you clarify what type of flowchart diagram you mean in point 2? Do you mean a flowchart diagram for the installation process?
