### 2020-09-24 WelbornGroup/ELECTRIC#19 @govarguz
@janash, I meant more a generic flowchart which spots where is Electric within molecular simulation environments. For example it works within the Molssi driver Interface, and requires simulation trajectories, and so on.
So that, one clearly identifies it as a post-processing/analysis tool within a simulation pipeline.
Does it answer your question?
