### 2020-09-24 WelbornGroup/ELECTRIC#19 @janash
Yes, I think so. I will draft a flowchart and we can iterate if necessary.
