### 2020-09-25 f87036a
Fix Tinker path in test_simulation.py
