+++ docs/index.rst
@@ -1,0 +1,21 @@
+.. ELECTRIC documentation master file, created by
+   sphinx-quickstart on Fri Sep 25 17:00:58 2020.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+ELECTRIC
+====================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   tutorial
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
