+++ docs/tutorial.rst
@@ -1,0 +1,15 @@
+Tutorial
+========
+
+This tutorial will walk you through using ELECTRIC to analyze the electric field in a small protein. This tutorial assumes you have ELECTRIC and MDI-enabled Tinker installed. If you don't, navigate to the [installation instructions]().
+
+The pdb code for this protein is [1l2y](https://www.rcsb.org/structure/1l2y), and you can see the structure below.
+
+.. moleculeView:: 
+    
+    data-pdb: 1l2y
+    data-backgroundcolor: white
+    width: 300px
+    height: 300px
+
+To follow along with this tutorial, you can clone the [tutorial repository](). Included in this repository is folder called `data`. The `data` directory has all of the 
