+++ docs/installation.rst
@@ -1,0 +1,4 @@
+Installation
+============
+
+Installation instructions will be here.
