### 2020-09-27 20b381f
update moleculeView directive
