+++ docs/_ext/moleculeView.py
@@ -32,8 +32,7 @@ def html_visit_output_node(self, node):
         "data-backgroundcolor": "white",
         "width": "400px",
         "height": "300px",
-        "data-style": "stick",
-        "data-style-color": None
+        "data-style": "stick"
     }
 
     if 'data-pdb' not in node.settings and 'data-href' not in node.settings:
@@ -53,16 +52,24 @@ def html_visit_output_node(self, node):
     elif 'data-href' in defaults:
         source_string = F'data-href={defaults["data-href"]}'
 
-    if defaults['data-style-color']:
-        data_style = { defaults['data-style'] : {"color": defaults['data-style-color']}}
-        data_style = json.dumps(data_style)
-    else:
-        data_style = defaults['data-style']
+    # Set data styles
+    data_style = ''
 
-    print(data_style)
+    # Process extra styles
+    style_keys = [k for k in defaults.keys() if 'style' in k]
+    select_keys = [k for k in defaults.keys() if 'select' in k]
+    
+    for i in range(len(style_keys)):
+        style_key = style_keys[i]
+        data_style += F" {style_key}={defaults[style_key]} "
+    
+    for i in range(len(select_keys)):
+        select_key = select_keys[i]
+        data_style += F" {select_key}={defaults[select_key]} "
 
-    self.body.append(F"""    
-         <center><div style="height: {defaults['height']}; width: {defaults['width']}; position: relative;" class='viewer_3Dmoljs' {source_string} data-backgroundcolor='0xffffff' data-style='{data_style}'></div></center>""")
+    show_string = F'<center><div style="height: {defaults["height"]}; width: {defaults["width"]}; position: relative;" class="viewer_3Dmoljs" {source_string} data-backgroundcolor="{defaults["data-backgroundcolor"]}" {data_style}></div></center>'
+
+    self.body.append(show_string)
 
 def html_depart_output_node(self, node):
     pass
