+++ docs/tutorial.rst
@@ -5,12 +5,13 @@ This tutorial will walk you through using ELECTRIC to analyze the electric field
 
 .. note::
     This tutorial will assume the following:
-        - You are able to run a molecular dynamics simulation using the Tinker software.
+        - You are able to run a molecular dynamics simulation using the Tinker software, or are familiar enough with molecular dynamics to follow along.
+        - You have installed ELECTRIC and MDI-enabled Tinker. If you have not, see the installation_ instructions.
         - You are able to download or clone a directory from git.
         - You are familiar with bash scripts.
 
 
-The pdb code for this protein is 1l2y_, and you can see the structure below.
+The pdb code for this protein is 1l2y_, and you can see the structure below. We have chosen a small protein for demonstrative purposes.
 
 .. moleculeView:: 
     
@@ -18,12 +19,181 @@ The pdb code for this protein is 1l2y_, and you can see the structure below.
     data-backgroundcolor: white
     width: 300px
     height: 300px
-    data-style: cartoon
-    data-style-color: spectrum
+    data-style: cartoon:color=spectrum
 
-To follow along with this tutorial, you can clone the `tutorial repository`_. Included in this repository is folder called `data`. The `data` directory has all of the data you will need for this tutorial.
+Running an ELECTRIC calculation
+###############################
 
-ELECTRIC is a post-processing analysis, meaning that you should first run your simulations using the AMOEBA forcefield and save the trajectory, and ELECTRIC performs analysis on that trajectory. We will be analyzing the included trajectory `1l2y_npt.arc`. This trajectory is a text file containing coordinates for a simulation snapshot we have run.
+Preparing Files
+----------------
+To follow along with this tutorial, clone the `tutorial repository`_. Included in this repository is folder called `data`. The `data` directory has all of the data you will need for this tutorial.
+
+ELECTRIC is a post-processing analysis, meaning that you should first run your simulations using the AMOEBA forcefield and save the trajectory. After you have a trajectory from Tinker simulation, use ELECTRIC to perform electric field analysis on that trajectory. We will be analyzing the trajectory included in the tutorial repository, `1l2y_npt.arc`. This trajectory is a text file containing coordinates for a simulation that has already been run.
+
+To get started with ELECTRIC, we will need to prepare our input files. We will need:
+    - a molecular dynamics trajectory
+    - a Tinker input file which does not have periodic boundaries or Ewald summation
+    - the forcefield parameter file
+    - a bash script file 
+
+We already have the molecular dynamics trajectory, so let's look at each of these additional files.
+
+Simulation Parameter File
+^^^^^^^^^^^^^^^^^^^^^^^^^
+This file contains the force field parameters for the simulation you have run. If you are using this software for analysis, use the same force field you used to run the molecular dynamics simulation. For this tutorial, the parameter file is `amoebabio18.prm`. We will need this for our Tinker input file (next section).
+
+Tinker input file
+^^^^^^^^^^^^^^^^^
+Next, we must prepare an input file which tells Tinker settings for our calculation. This input file should be a modified version of the one which you used to run your initial simulation. Consider the input file, `tinker.key` used to obtain this trajectory:
+
+.. code-block:: text
+
+    parameters amoebabio18.prm 
+    openmp-threads 16
+
+    a-axis 50.00 
+    b-axis 50.00
+    c-axis 50.00
+
+    polar-eps 0.000010
+    polar-predict
+    polarization mutual
+
+
+    cutoff 10.0
+    ewald
+    neighbor-list
+    integrator beeman
+
+    thermostat nose-hoover
+    barostat nose-hoover
+
[…]
