+++ docs/_ext/moleculeView.py
@@ -1,5 +1,3 @@
-import json
-
 from docutils import nodes
 from docutils.parsers.rst import Directive
 
