### 2020-09-28 475a5bb
change tutorial for shorter simulation trajectory
