+++ docs/tutorial.rst
@@ -125,7 +125,7 @@ You use ELECTRIC from the command line. Consider the following bash script provi
     ${TINKER_LOC} 1l2y -k no_ewald.key -mdi "-role ENGINE -name NO_EWALD -method TCP -port 8022 -hostname localhost"  10 1.0 0.002 2 300.00 > no_ewald.log &
 
     #launch driver
-    python ${DRIVER_LOC} -snap 1l2y_npt.arc -probes "93 94" -mdi "-role DRIVER -name driver -method TCP -port 8022" --byres 1l2y_solvated.pdb  --equil 160 --stride 2 &
+    python ${DRIVER_LOC} -snap 1l2y_npt.arc -probes "93 94" -mdi "-role DRIVER -name driver -method TCP -port 8022" --byres 1l2y_solvated.pdb  --equil 140 --stride 2 &
 
     wait
 
@@ -147,14 +147,14 @@ In the next command (line `21`), we launch ELECTRIC.
 
 .. code-block:: bash   
 
-    python ${DRIVER_LOC} -snap 1l2y_npt.arc -probes "78 93 94"  -mdi "-role DRIVER -name driver -method TCP -port 8022" --byres 1l2y_solvated.pdb  --equil 160 --stride 2 &
+    python ${DRIVER_LOC} -snap 1l2y_npt.arc -probes "78 93 94"  -mdi "-role DRIVER -name driver -method TCP -port 8022" --byres 1l2y_solvated.pdb  --equil 140 --stride 2 &
 
 Here, we first give the location of our ELECTRIC driver. We indicate our trajectory file using the `-snap` argument with the filename to analyze, followed by MDI options.
 
 Probe Atoms 
 ++++++++++++
 
-To run an ELECTRIC calculation, you must give the indices of your probe atoms. You should obtain these from the `xyz` file you use to launch MDI-enabled Tinker (note that this could be different than your pdb file if hydrogens are handled differently). Note that the index you use here should match the number given in the first column of your xyz file. The projection of the electric field at the midpoint of these two atoms will be reported for each analyzed frame. If you indicate more than two probes, all pairwise fields will be reported (ie, if using "78 93 94", you will get "78 and 93", "78 and 94" and "93 and 94"). You can see the atoms we have chosen as probes highlighted below:
+To run an ELECTRIC calculation, you must give the indices of your probe atoms. The probe atoms are the atoms where you choose to analyze the electric field. You should obtain the number of the probe atoms from the `xyz` file you use to launch MDI-enabled Tinker (note that this could be different than your pdb file if hydrogens are handled differently). Note that the index you use here should match the number given in the first column of your xyz file. The projection of the electric field at the midpoint of these two atoms will be reported for each analyzed frame. If you indicate more than two probes, all pairwise fields will be reported (ie, if using "78 93 94", you will get "78 and 93", "78 and 94" and "93 and 94"). You can see the atoms we have chosen as probes highlighted below:
 
 .. moleculeView:: 
     
@@ -166,7 +166,7 @@ To run an ELECTRIC calculation, you must give the indices of your probe atoms. Y
     data-select1: serial:78,93,94
     data-style1: sphere
 
-The argument `--byres` gives information about how we would like the electric field reported. When we use the `--byres` argument, it should be followed by a pdb which contains residue information for the system you are studying. When using this argument, electric field contributions from each residue will be reported. Other options are `--byatom` top report electric field contributions from each atom, and `--bymol` to report electric field contributions from each molecule. 
[…]
