### 2020-09-28 13f43c1
try adding a github action to build docs
