+++ paper.md
@@ -31,7 +31,7 @@ Polarizable force fields have changed the landscape of biomolecular simulation, 
 
 Improved electrostatics with AMOEBA give a unique opportunity to accurately compute electric fields, powerful metrics of catalytic activity in enzymes and other systems.[@doi:10.1021/jacs.6b12265; @doi:10.1021/acscatal.7b03151; @natcat] Electric fields projected onto specific bonds report on the effect of the surroundings (interacting via coulombic interactions, solvent effects, hydrogen bonding or other forces, all mostly electrostatic in nature) on the flow of electrons along these bonds. Therefore, projected electric fields are correlated to the probability of breaking these bonds, making them a useful probe of chemical reactivity.
 
-`ELECTRIC` [@electric] is a MolSSI Driver Interface (MDI) [@mdi_repo; @barnes_taylor_arnold_2020_3659285] driver that utilizes Tinker [@doi:10.1021/acs.jctc.8b00529] to analyze specific components of electric fields that are modeled using the AMOEBA force field.  `ELECTRIC` parses Tinker trajectories and orchestrates additional Tinker calculations in order to project components of the electric fields onto user-defined bonds (specified by two atoms). It outputs the field in MV/cm, which is the sum of the direct field (from permanent electrostatics) and the induced field (from the induce dipole term), projected onto the bond unit vector (i.e. normalized by the bond length). `ELECTRIC` enables splitting of the total field into contributions from different components of the system, by molecules or by residues as specified in a reference PDB file. In summary, `ELECTRIC` was designed to improve quantitative system characterization via the computation of electric fields with user-friendly processing tools of Tinker-AMOEBA simulations.
+`ELECTRIC` [@electric] is a MolSSI Driver Interface (MDI) [@mdi_repo; @barnes_taylor_arnold_2020_3659285] driver that utilizes Tinker [@doi:10.1021/acs.jctc.8b00529] to analyze specific components of electric fields that are modeled using the AMOEBA force field.  `ELECTRIC` parses Tinker trajectories and orchestrates additional Tinker calculations in order to project components of the electric fields onto user-defined bonds (specified by two atoms). It outputs the field in MV/cm, which is the sum of the direct field (from permanent electrostatics) and the induced field (from the induce dipole term), projected onto the bond unit vector (i.e. normalized by the bond length). `ELECTRIC` enables splitting of the total field into contributions from different components of the system, by molecules or by residues as specified in a reference PDB file. In summary, `ELECTRIC` was designed to expand quantitative system characterization via the computation of electric fields with user-friendly processing tools of Tinker-AMOEBA simulations.
 
 # Mathematics
 The electric field at atom $i$, $\vec{E}^i$, has components defined as:
