### 2020-09-28 #2576 @valeriewelborn
@richardjgowers The modifications to Tinker that enable the MDI interface used by ELECTRIC will be hosted by MolSSI at https://github.com/MolSSI-MDI/Tinker/commits/mdi-ef as part of the MolSSI MDI project. These are the Tinker changes that are part of this work. Further, we have added the MolSSI-MDI-enabled Tinker as a submodule of ELECTRIC, as suggested in issue WelbornGroup/ELECTRIC#17. WelbornGroup/Tinker_ELECTRIC has been deleted to avoid confusion.
