### 2020-09-28 6f1e1f7
add directive to use datatables
