+++ docs/tutorial.rst
@@ -32,7 +32,7 @@ ELECTRIC is a post-processing analysis, meaning that you should first run your s
 
 To get started with ELECTRIC, we will need to prepare our input files. We will need:
     - a molecular dynamics trajectory
-    - a Tinker input file which does not have periodic boundaries or Ewald summation
+    - a Tinker input file (usually called a key file) which does not have settings for periodic boundaries or Ewald summation
     - the forcefield parameter file
     - a bash script file 
 
@@ -44,9 +44,10 @@ This file contains the force field parameters for the simulation you have run. I
 
 Tinker input file
 ^^^^^^^^^^^^^^^^^
-Next, we must prepare an input file which tells Tinker settings for our calculation. This input file should be a modified version of the one which you used to run your initial simulation. Consider the input file, `tinker.key` used to obtain this trajectory:
+Next, we must prepare an input file which tells Tinker settings for our calculation. This input file should be a modified version of the one which you used to run your initial simulation. Consider the input file, `tinker.key` used to obtain this trajectory. The parameter file in the previous step is given on line 2.
 
 .. code-block:: text
+    :linenos:
 
     parameters amoebabio18.prm 
     openmp-threads 16
@@ -72,11 +73,11 @@ Next, we must prepare an input file which tells Tinker settings for our calculat
     printout 1000
 
 
-The input file used for this simulation uses both periodic boundaries and an Ewald summation for electrostatics. During a Tinker simulation using AMOEBA, electric fields are calculated in order to calculate the induced dipoles at each step. In order to get electric field contributions from specific residues, we must calculate the electric field using the real space interactions only (no periodic boundaries or Ewald). 
+The input file used for this simulation uses periodic boundaries and an Ewald summation for electrostatics. During a Tinker simulation using AMOEBA, electric fields are calculated in order to calculate the induced dipoles at each step. In order to get electric field contributions from specific residues, we must calculate the electric field using the real space interactions only (no periodic boundaries or Ewald). 
 
 Remove settings related to cutoffs (`cutoff` keyword), periodic boundaries (`a-axis`, `b-axis`, `c-axis`) and Ewald summation (`ewald`). You can also remove settings having to do with neighbor lists (`neighbor-list`), as they are not needed and can cause an error for this calculation if included.
 
-The modifed input file for ELECTRIC is given below. Save this file with the name `no_ewald.key`.
+The modifed input file for ELECTRIC is given below. This file is saved in the data directory with the name `noewald.key`.
 
 .. code-block:: text
 
@@ -96,9 +97,9 @@ The modifed input file for ELECTRIC is given below. Save this file with the name
     printout 100
 
 
-Bash script
-^^^^^^^^^^^
-This file will launch MDI-enabled Tinker and the ELECTRIC driver. When you run an ELECTRIC analysis, ELECTRIC parses your given trajectory and options and sends snapshots to Tinker. The MDI-enabled version of Tinker then calculates the electric field information for that snapshot, as would usually be calculated during a molecular dynamics simulation. 
+Bash script - run_analysis.sh
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+When you run analysis uisng ELECTRIC, ELECTRIC parses your given trajectory sends snapshots to Tinker for electric field calculation. The MDI-enabled version of Tinker then calculates the electric field information for that snapshot, as would usually be calculated during a molecular dynamics simulation. 
 
 You use ELECTRIC from the command line. Consider the following bash script provided for analysis, `run_analysis.sh`. We will explain this script in detail.
 
@@ -129,7 +130,11 @@ You use ELECTRIC from the command line. Consider the following bash script provi
 
     wait
 
[…]
