+++ docs/usage.rst
@@ -1,70 +1,17 @@
-[![Build Status](https://travis-ci.com/WelbornGroup/ELECTRIC.svg?branch=master)](https://travis-ci.com/WelbornGroup/ELECTRIC)
-[![codecov](https://codecov.io/gh/WelbornGroup/ELECTRIC/branch/master/graph/badge.svg)](https://codecov.io/gh/WelbornGroup/ELECTRIC)
-[![Language grade: Python](https://img.shields.io/lgtm/grade/python/g/WelbornGroup/ELECTRIC.svg?logo=lgtm&logoWidth=18)](https://lgtm.com/projects/g/WelbornGroup/ELECTRIC/context:python)
+Usage
+=====
 
-# ELECTRIC
+Information Flow
+----------------
 
-The `ELECTRIC` driver has been written by Jessica Nash and Taylor Barnes, from the [MolSSI](https://molssi.org/). 
+ELECTRIC is a post-processing tool for simulations running with the AMOEBA polarizable force field using the Tinker software package.
+
+.. image:: images/inputs_and_outputs.svg
+   :width: 600
 
 
-## Overview
-
-This repository contains a driver that uses the [MolSSI Driver Interface](https://github.com/MolSSI-MDI/MDI_Library) to perform electric field analysis of [Tinker](https://dasher.wustl.edu/tinker/) trajectories which use the AMOEBA forcefield. This currently works as a post-processing tool, meaning that you run simulations as normal using Tinker, then analyze the trajectories using
-MDI-enabled Tinker and this driver.
-
-Using this tool, you can calculate the electric field along a bond or between atoms due to molecules or residues in the system.
-
-### Compiling MDI-Tinker and ELECTRIC
-
-Installation of ELECTRIC and MDI-enabled Tinker are bundled in one convenient build script. 
-
-To install ELECTRIC and MDI-enabled Tinker, you should have cmake and a fortran compiler installed. Then, you can download and build ELECTRIC and MDI-enabled Tinker using the following command in your terminal. Make sure you are in the directory where you want your ELECTRIC driver to be. You should note this location, because you will need to specify the path to some files built during this process in order to perform analysis.
-
-```
-git clone --recurse-submodules https://github.com/WelbornGroup/ELECTRIC.git
-cd ELECTRIC
-./build.sh
-```
-
-This will download and build ELECTRIC and MDI-enabled Tinker. 
-
-Upon successfull building, you will have the ELECTRIC driver in ELECTRIC/ELECTRIC/ELECTRIC.py, and the needed Tinker executable (dynamic.x) in ELECTRIC/modules/Tinker/build/tinker/source/dynamic.x . The location of these files can be found in text files in ELECTRIC/test/locations/ELECTRIC and ELECTRIC/test/locations/Tinker_ELECTRIC. You will need these for using ELECTRIC.
-
-### Python Dependencies
-
-In order to run ELECTRIC, you will need to be in a python environment which has numpy and pandas installed. We recommend installing these packages in a conda environment created for ELECTRIC analysis.
-
-``` 
-conda install -c conda-forge numpy pandas
-```
-
-## Testing
-
-You can now run a quick test of the driver by changing directory to the `ELECTRIC/test/bench5` directory and running the `tcp.sh` script:
-
-    ./tcp.sh
-
-This script will run a short Tinker dynamics simulation that includes periodic boundary conditions. This command is on line 20 of the provided file. This is a standard Tinker call, as you would normally run a simulation. If you are performing post processing on a simulation, you will not use this line.
-
-```
-${TINKER_LOC} bench5 -k bench5.key 10 1.0 0.001999 2 300.00 > Dynamics.log
-```
-
[…]
