### 2020-09-29 69b4e03
switch to using pandas testing function instead of manual comparison
