### 2020-09-29 19c448a
update stride to consider equilibration length
