+++ ELECTRIC/ELECTRIC.py
@@ -159,8 +159,7 @@ if __name__ == "__main__":
 
     optional.add_argument("--equil", help='''The number of frames to skip performing analysis on
                 at the beginning of the trajectory file (given by the -snap argument)
-                For example, using --equil 50 will result in the first 50 frames of the trajectory
-                being skipped.''',
+                For example, using --equil 50 will result in analysis starting on frame 50 of the trajectory.''',
                 type=int, default=0)
 
     optional.add_argument("--stride", help='''The number of frames to skip between
@@ -333,7 +332,7 @@ if __name__ == "__main__":
         skiprows=skip_line, index_col=None), 1):
 
         if snap_num > equil - 1:
-            if snap_num % stride == 0:
+            if (snap_num - equil) % stride == 0:
 
                 icomm = itask % nengines
                 itask_to_snap_num[itask] = snap_num
