### 2020-09-29 cf80f4c
change so that analysis starts after equil
