### 2020-09-29 4252fcc
update docs to autodocument ELECTRIC cl arguments
