+++ docs/conf.py
@@ -58,10 +58,10 @@ html_theme = "sphinx_rtd_theme"
 # Custom directives
 sys.path.append(os.path.abspath("./_ext"))
 
-extensions = ["moleculeView", "dataTables", "sphinx_rtd_theme",]
+extensions = ["moleculeView", "dataTables", "sphinx_rtd_theme", 'sphinxarg.ext']
 
 def setup(app):
     app.add_js_file('https://3dmol.org/build/3Dmol-min.js')
     app.add_js_file('https://code.jquery.com/jquery-3.5.1.min.js')
     app.add_js_file('https://cdn.datatables.net/1.10.22/js/jquery.dataTables.js')
-    app.add_css_file('https://cdn.datatables.net/1.10.22/css/jquery.dataTables.css')
+    #app.add_css_file('https://cdn.datatables.net/1.10.22/css/jquery.dataTables.css')
