+++ docs/requirements.txt
@@ -1,1 +1,2 @@
 sphinx_rtd_theme
+pip install sphinx-argparse
