### 2020-09-29 dfd94d2
doc update- synatx and tutorial
