### 2020-09-29 4d23446
update .github for master branch and to point to main repo
