+++ .github/workflows/sphinx_build.yml
@@ -1,13 +1,15 @@
 name: CI
 
-on:
+on: 
   push:
     branches:    
-      - docs
+      - master
+  pull_request:
+    branches:
+      - master
 
 jobs:
   build:
-
     runs-on: ubuntu-latest
 
     steps:
@@ -21,7 +23,7 @@ jobs:
     # ===============================
     - name: Commit documentation changes
       run: |
-        git clone https://github.com/janash/ELECTRIC --branch gh-pages --single-branch gh-pages
+        git clone https://github.com/WelbornGroup/ELECTRIC --branch gh-pages --single-branch gh-pages
         cp -r docs/_build/html/* gh-pages/
         cd gh-pages
         git config --local user.email "action@github.com"
@@ -30,7 +32,9 @@ jobs:
         git commit -m "Update documentation" -a || true
         # The above command will fail if no changes were present, so we ignore
         # that.
+    # Push if this is a push to master.
     - name: Push changes
+      if: ${{ github.event_name == 'pull_request' }}
       uses: ad-m/github-push-action@master
       with:
         branch: gh-pages
