### 2020-09-29 9547997
fix gha to trigger docs build on push to master
