+++ _sources/usage.rst.txt
@@ -1,0 +1,114 @@
+Usage
+=====
+
+Information Flow
+----------------
+
+ELECTRIC is a post-processing tool for simulations running with the AMOEBA polarizable force field using the Tinker software package.
+
+.. image:: images/inputs_and_outputs.svg
+   :width: 600
+
+
+Procedure
+---------
+
+In general, running a calculation with the driver requires the following steps:
+
+1. **Run a dynamics simulation with Tinker.**  
+This simulation should be run with periodic boundary conditions (if desired), and should print snapshots of its results to a single file (i.e., :code:`coordinates.arc`).
+If each snapshot was instead written to a different file (i.e., :code:`coordinates.001`, :code:`coordinates.002`, etc.) then you may concatenate them into a single file.
+
+2. **Create a new Tinker keyfile.**   
+This keyfile should be identical to the one used in Step 1, except that it **must not** include periodic boundary conditions and **must not** use an Ewald summation. This means that in the :code:`.key` file for running the driver, you should not have an :code:`a-axis` keyword, or keywords related to Ewald.
+
+3. **Launch one (or more; see the `-nengines` option below) instance(s) of Tinker as an MDI engine, using the keyfile created in Step 2.**  
+This is done in the same way you launch a normal Tinker simulation (by launching the :code:`dynamic.x` executable) except that the `-mdi` command-line option is added. However, it is **very important** that the reference coordinates you use do not have periodic boundary information. So, if when you originally ran the simulation you started it with a snapshot from a previous simulation run, make sure to create a new snapshot to launch the simulation from which does not include box information on line 2.
+
+The argument to the :code:`-mdi` command-line option details how Tinker should connect to the driver; its possible arguments are described in the `MDI documentation`_ .
+When in doubt, we recommend doing :code:`-mdi "-role ENGINE -name NO_EWALD -method TCP -port 8021 -hostname localhost"`
+When run as an engine, Tinker should be launched in the background; this is done by adding an ampersand (:code:`&`) at the end of the launch line.
+
+4. **Launch the driver**
+The driver accepts a variety of command-line options, which are described in detail below.
+One possible launch command would be:
+
+.. code-block:: bash
+
+    `python ${DRIVER_LOC} -probes "1 2 10" -snap coordinates.arc -mdi "-role DRIVER -name driver -method TCP -port 8021" --byres ke15.pdb --equil 51 -nengines 15 &`
+
+where `DRIVER_LOC` is the path to ELECTRIC.py which you set during the configuration step. See the section :ref:`electric settings` for a detailed explanation of command line arguments for ELECTRIC.
+
+The output will be written to `proj_totfield.csv`.
+
+It is useful to write a script that performs Steps 3 and 4, especially if the calculations are intended to be run on a shared cluster.
+Such a script might look like:
+
+.. code-block:: bash
+
+    # location of required codes
+    DRIVER_LOC=$(cat ../locations/ELECTRIC)
+    TINKER_LOC=$(cat ../locations/Tinker_ELECTRIC)
+
+    # number of instances of Tinker to run as an engine
+    nengines=18
+
+    # set the number of threads used by each code
+    export OMP_NUM_THREADS=1
+
+    # launch Tinker as an engine
+    for i in $( eval echo {1..$nengines} )
+    do
+    ${TINKER_LOC} coordinates.in -k no_ewald.key -mdi "-role ENGINE -name NO_EWALD -method TCP -port 8021 -hostname localhost" 10 1.0 1.0 2 300 > no_ewald${i}.log &
+    done
+
+    # launch the driver
[…]
