### 2020-09-29 WelbornGroup/ELECTRIC#19 @janash
Hi @govarguz, I've attempted to address your points. Most of these seemed like they could be fixed through improved documentation, so please see the new GitHub pages for this software : https://welborngroup.github.io/ELECTRIC/

1. and 3. - The tutorial on the GitHub pages site gives a tutorial for running an analysis on a protein MD simulation using the `--byres` option - https://welborngroup.github.io/ELECTRIC/tutorial.html . We also felt that this was a good response to point 3. ELECTRIC is a python script, but runs the compiled Fortran program MDI-Tinker in the background. This seems too computationally expensive for a Binder session. ELECTRIC is also intended to be used as command line tool, in a script similar to what you would see on the Usage page (https://welborngroup.github.io/ELECTRIC/usage.html). We felt this step-by-step tutorial was a good alternative to a Jupyter notebook.

2. We have added flowchart diagram on the Usage page (https://welborngroup.github.io/ELECTRIC/usage.html) in attempt to show where ELECTRIC fits in relative to running your MD simulation.

Other points of improvement - We have also greatly simplified the installation process, thanks to a suggestion from @amandadumi
