### 2020-10-02 WelbornGroup/ELECTRIC#18 @amandadumi
Looks great! Thanks!
