### 2020-10-02 WelbornGroup/ELECTRIC#25 @amandadumi
JOSS: Code review
The code seems very clean and concise, very good work.  Here are a few notes.

Electric
-----
## File: Electric.py
- [ ] a general note is to adopt a formatting standard and list this in the contributing, such as autopep8 or black. This would also belong in the CONTRIBUTING.md, or wherever you choose to discuss community contributions as discussed in #17. This may already be done on your end, but just worth a mention for the community aspect.
### function:  `mdi_checks`:
  - [ ] need to define input arguments in documentation.
### function: `collect_task`:
- [ ] define input/output argument data types and definitions.
- [ ] npoles from_fragment seem to be used by the function collect_task, though instead of being passed as a parameter to the function it is inferred from the definition in main, which is hard to follow and could lead to problems in development and a tricker time at debugging for other contributors.
### function: main
- [ ] capitalization of Tinker on line 241
- [ ] line 174: multipole misspelled.

## File: Util.py
### function:  `create_parser`
  - [ ] Consistency among optional argument dashes. More specifically `-nengines` only required one dash where as every other requires two.

### function: `print_info`
  - [ ] adding a description for the input argument for consistency.

### function: `index fragment`
 - [ ] only one of the return arguments is explained.
 - [ ] "specified" is misspelled in the comments
