### 2020-10-02 WelbornGroup/ELECTRIC#17 @amandadumi
Hello, just checking in on this issue to say that the module implementation looks great! I'm glad that worked out. I've listed some more specifics in #25 on util.py and electric.py as the changes to documentation changes ended up being within the function definitions in the code.
