### 2020-10-06 #2576 @amandadumi
I wanted to add to this thread that https://github.com/WelbornGroup/ELECTRIC/issues/25 reviews the code within the ELECTRIC project. I plan on going through the mdi-ef code this evening. The simplification of mdi-ef being hosted by MolSSI is great, easier for users to open issues that pertain to the MDI changes specifically and hopefully easier for this authors to maintain as they work toward integrating this tool into the main Tinker project.
