### 2020-10-06 4c9dbe8
clean up docstrings, connect_to_engines function
