### 2020-10-06 4497cc4
run black on docs files
