+++ docs/_ext/moleculeView.py
@@ -11,22 +11,24 @@ from sphinx.util.docutils import SphinxDirective
 class view(nodes.General, nodes.Element):
     pass
 
+
 class MoleculeView(SphinxDirective):
 
     has_content = True
 
     def run(self):
-        view_id = F'viewport-{self.env.new_serialno("view")}' 
+        view_id = f'viewport-{self.env.new_serialno("view")}'
         settings = {}
 
         for content in self.content:
-            split_content = content.split(':', 1)
-            settings[split_content[0] ] = split_content[1]
+            split_content = content.split(":", 1)
+            settings[split_content[0]] = split_content[1]
 
         node = view()
         node.view_id = view_id
         node.settings = settings
         return [node]
+
 
 def html_visit_output_node(self, node):
 
@@ -34,44 +36,45 @@ def html_visit_output_node(self, node):
         "data-backgroundcolor": "white",
         "width": "400px",
         "height": "300px",
-        "data-style": "stick"
+        "data-style": "stick",
     }
 
-    if 'data-pdb' not in node.settings and 'data-href' not in node.settings:
+    if "data-pdb" not in node.settings and "data-href" not in node.settings:
         raise KeyError("You must specify a path to a pdb or file to visualize.")
-    
-    if 'data-pdb' and 'data-href' in node.settings:
+
+    if "data-pdb" and "data-href" in node.settings:
         raise ValueError("You should only specify one molecule file to visualize.")
 
     for k, v in node.settings.items():
         k = k.strip()
         defaults[k] = v.strip()
 
-    source_string = ''
+    source_string = ""
 
-    if 'data-pdb' in defaults:
-        source_string = F'data-pdb={defaults["data-pdb"]}'
-    elif 'data-href' in defaults:
-        source_string = F'data-href={defaults["data-href"]}'
+    if "data-pdb" in defaults:
+        source_string = f'data-pdb={defaults["data-pdb"]}'
+    elif "data-href" in defaults:
+        source_string = f'data-href={defaults["data-href"]}'
 
     # Set data styles
-    data_style = ''
+    data_style = ""
 
     # Process extra styles
-    style_keys = [k for k in defaults.keys() if 'style' in k]
-    select_keys = [k for k in defaults.keys() if 'select' in k]
-    
+    style_keys = [k for k in defaults.keys() if "style" in k]
+    select_keys = [k for k in defaults.keys() if "select" in k]
+
     for i in range(len(style_keys)):
         style_key = style_keys[i]
-        data_style += F" {style_key}={defaults[style_key]} "
-    
+        data_style += f" {style_key}={defaults[style_key]} "
+
     for i in range(len(select_keys)):
         select_key = select_keys[i]
-        data_style += F" {select_key}={defaults[select_key]} "
+        data_style += f" {select_key}={defaults[select_key]} "
 
-    show_string = F'<center><div style="height: {defaults["height"]}; width: {defaults["width"]}; position: relative;" class="viewer_3Dmoljs" {source_string} data-backgroundcolor="{defaults["data-backgroundcolor"]}" {data_style}></div></center>'
+    show_string = f'<center><div style="height: {defaults["height"]}; width: {defaults["width"]}; position: relative;" class="viewer_3Dmoljs" {source_string} data-backgroundcolor="{defaults["data-backgroundcolor"]}" {data_style}></div></center>'
 
     self.body.append(show_string)
+
 
 def html_depart_output_node(self, node):
     pass
@@ -81,17 +84,13 @@ def setup(app):
 
     app.add_node(
         view,
-        html=(
-            html_visit_output_node,
-            html_depart_output_node
-        ),
+        html=(html_visit_output_node, html_depart_output_node),
     )
 
[…]
