### 2020-10-06 ad97422
clean up ELECTRIC.py
