+++ ELECTRIC/ELECTRIC.py
@@ -11,7 +11,9 @@ import mdi.MDI_Library as mdi
 
 try:
     from mpi4py import MPI
+
     use_mpi4py = True
+
 except ImportError:
     use_mpi4py = False
 
@@ -20,6 +22,7 @@ if use_mpi4py:
     mpi_world = MPI.COMM_WORLD
 else:
     mpi_world = None
+
 
 def connect_to_engines(nengines):
     """
@@ -53,12 +56,13 @@ def connect_to_engines(nengines):
         # Determine the name of the engine
         mdi.MDI_Send_Command("<NAME", comm)
         name = mdi.MDI_Recv(mdi.MDI_NAME_LENGTH, mdi.MDI_CHAR, comm)
-        print(F"Engine name: {name}")
+        print(f"Engine name: {name}")
 
         if name != "NO_EWALD":
             raise Exception("Unrecognized engine name", name)
 
     return engine_comm
+
 
 def collect_task(comm, npoles, snapshot_coords, snap_num, atoms_pole_numbers, output):
     """
@@ -77,7 +81,7 @@ def collect_task(comm, npoles, snapshot_coords, snap_num, atoms_pole_numbers, ou
 
     snap_num : int
         The snapshot associated with this task.
-    
+
     atoms_pole_numbers : list
         A multidimensional list where each element gives the pole indices in that fragment.
 
@@ -90,63 +94,66 @@ def collect_task(comm, npoles, snapshot_coords, snap_num, atoms_pole_numbers, ou
         The aggregated data from all tasks, including the data collected by this function.
     """
 
-    mdi.MDI_Recv(3*npoles*len(probes), mdi.MDI_DOUBLE, comm, buf=dfield)
+    mdi.MDI_Recv(3 * npoles * len(probes), mdi.MDI_DOUBLE, comm, buf=dfield)
 
     # Get the pairwise UFIELD
-    ufield = np.zeros((len(probes),npoles,3))
+    ufield = np.zeros((len(probes), npoles, 3))
     mdi.MDI_Send_Command("<UFIELD", comm)
-    mdi.MDI_Recv(3*npoles*len(probes), mdi.MDI_DOUBLE, comm, buf=ufield)
+    mdi.MDI_Recv(3 * npoles * len(probes), mdi.MDI_DOUBLE, comm, buf=ufield)
 
     # Sum the appropriate values
 
-    columns = ['Probe Atom', 'Probe Coordinates']
-    columns += [F'{by_type} {x}' for x in from_fragment]
+    columns = ["Probe Atom", "Probe Coordinates"]
+    columns += [f"{by_type} {x}" for x in from_fragment]
     dfield_df = pd.DataFrame(columns=columns, index=range(len(probes)))
     ufield_df = pd.DataFrame(columns=columns, index=range(len(probes)))
     totfield_df = pd.DataFrame(columns=columns, index=range(len(probes)))
 
     # Get sum at each probe from fragment.
     for i in range(len(probes)):
-        dfield_df.loc[i, 'Probe Atom'] = probes[i]
-        dfield_df.loc[i, 'Probe Coordinates'] = snapshot_coords[probes[i]-1]
-        ufield_df.loc[i, 'Probe Atom'] = probes[i]
-        ufield_df.loc[i, 'Probe Coordinates'] = snapshot_coords[probes[i]-1]
-        totfield_df.loc[i, 'Probe Atom'] = probes[i]
-        totfield_df.loc[i, 'Probe Coordinates'] = snapshot_coords[probes[i]-1]
+        dfield_df.loc[i, "Probe Atom"] = probes[i]
+        dfield_df.loc[i, "Probe Coordinates"] = snapshot_coords[probes[i] - 1]
[…]
