### 2020-10-06 f9b7d24
clean up util.py
