### 2020-10-06 9a3f85d
run black - residue_report.py
