+++ ELECTRIC/residue_report.py
@@ -10,7 +10,7 @@ import argparse
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
 
-    parser.add_argument('pdb_file', help='The pdb file to analyze.')
+    parser.add_argument("pdb_file", help="The pdb file to analyze.")
 
     args = parser.parse_args()
 
