### 2020-10-06 27c1f4f
add contributing guidelines and PR template
