### 2020-10-07 c7a509b
fix call to collect_task
