### 2020-10-07 WelbornGroup/ELECTRIC#19 @govarguz
@janash
I see. In my opinion ELECTRIC is ready for publication at JOSS
