### 2020-10-07 #2576 @govarguz
@richardjgowers, ELECTRIC is ready for publication at JOSS
