### 2020-10-07 #2576 @amandadumi
The ELECTRIC project and the MDI-enabled Tinker look great. The authors have addressed any concerns and I agree that ELECTRIC is ready for publication at JOSS!
