### 2020-10-09 #2576 @richardjgowers
I need to go over this one last time this weekend then probably it’ll get
published :) thanks for your work on this!

On Fri, Oct 9, 2020 at 20:11, Valerie Welborn <notifications@github.com>
wrote:

> Thank you @amandadumi <https://github.com/amandadumi> and @govarguz
> <https://github.com/govarguz> for reviewing our paper. @richardjgowers
> <https://github.com/richardjgowers>, is there anything else you need from
> us?
>
