### 2020-10-12 d85d560
adds a paragraph about usage procedure to the paper
