+++ paper.md
@@ -33,6 +33,8 @@ Improved electrostatics with AMOEBA give a unique opportunity to accurately comp
 
 `ELECTRIC` [@electric] is a MolSSI Driver Interface (MDI) [@mdi_repo; @barnes_taylor_arnold_2020_3659285] driver that utilizes Tinker [@doi:10.1021/acs.jctc.8b00529] to analyze specific components of electric fields that are modeled using the AMOEBA force field.  `ELECTRIC` parses Tinker trajectories and orchestrates additional Tinker calculations in order to project components of the electric fields onto user-defined bonds (specified by two atoms). It outputs the field in MV/cm, which is the sum of the direct field (from permanent electrostatics) and the induced field (from the induce dipole term), projected onto the bond unit vector (i.e. normalized by the bond length). `ELECTRIC` enables splitting of the total field into contributions from different components of the system, by molecules or by residues as specified in a reference PDB file. In summary, `ELECTRIC` was designed to expand quantitative system characterization via the computation of electric fields with user-friendly processing tools of Tinker-AMOEBA simulations.
 
+In practice, the user needs a Tinker trajectory file (i.e. `filename.arc`), a Tinker input file (i.e. `tinker.key`) stripped from any keywords corresponding to periodic boundary conditions and a Tinker reference snapshot (`filename.xyz` - this can be the first frame in `filename.arc`) where the dimensions of the box, usually printed on line 2, has been deleted. Tinker is then launched as an MDI engine before the `ELECTRIC` driver. Options regarding the electric fields calculations, such as probe index number, number of frames at the beginning of the trajectory file to skip (equilibration procedure), reference `PDB` file, etc. are specified as command line arguments when launching the driver. A complete usage procedure is provided in the `README` file. 
+
 # Mathematics
 The electric field at atom $i$, $\vec{E}^i$, has components defined as:
 \begin{equation}
