### 2020-10-21 #2576 @valeriewelborn
Thank you @richardjgowers. The version tag is v1.0 and the DOI of the archived version is 10.5281/zenodo.4116258
