### 2020-10-25 #2576 @kyleniemeyer
Hi @valeriewelborn, I am the EIC on duty doing some final checks before publishing your article. There are some minor issues in the paper, could you address these?

- [ ] Most of the citations appear after the period in their sentences—could you move the citation commands to include them in the sentences?
- [ ] Could you put commas after "i.e." throughout? I noticed particularly in the third and fourth paragraphs.
- [ ] Your article is missing an explicit [Statement of Need section](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), which we have begun requiring for all articles. Could you please add this? It doesn't need to be too long, and you can likely use/reorganize some of what is already in your article.
