+++ paper.md
@@ -27,13 +27,16 @@ bibliography: paper.bib
 
 # Summary
 
-Polarizable force fields have changed the landscape of biomolecular simulation, mostly by featuring improved electrostatic potential energy terms.[@doi:10.1146/annurev-biophys-070317-033349] These novel energy functions allow environment-driven changes in charge distribution, which yield simulations with improved geometries and molecular properties. In particular, the AMOEBA polarizable force field exhibits two fundamental changes compared to more traditional fixed charge force fields.[@doi:10.1021/ct4003702; @doi:10.1021/acs.jctc.7b01169] The first one relates to permanent electrostatics, expressed in AMOEBA in terms of atomic multipoles (truncated at quadrupoles) that account for anisotropy in the computed charge distributions. The second one represents polarizability through an induced dipole term that can respond to the chemical environment. These modified terms make the AMOEBA force field more physically grounded than other force fields and is the basis for more realistic simulations of biomolecular systems.
+Polarizable force fields have changed the landscape of biomolecular simulation, mostly by featuring improved electrostatic potential energy terms[@doi:10.1146/annurev-biophys-070317-033349]. These novel energy functions allow environment-driven changes in charge distribution, which yield simulations with improved geometries and molecular properties. In particular, the AMOEBA polarizable force field exhibits two fundamental changes compared to more traditional fixed charge force fields[@doi:10.1021/ct4003702; @doi:10.1021/acs.jctc.7b01169]. The first one relates to permanent electrostatics, expressed in AMOEBA in terms of atomic multipoles (truncated at quadrupoles) that account for anisotropy in the computed charge distributions. The second one represents polarizability through an induced dipole term that can respond to the chemical environment. These modified terms make the AMOEBA force field more physically grounded than other force fields and is the basis for more realistic simulations of biomolecular systems.
 
-Improved electrostatics with AMOEBA give a unique opportunity to accurately compute electric fields, powerful metrics of catalytic activity in enzymes and other systems.[@doi:10.1021/jacs.6b12265; @doi:10.1021/acscatal.7b03151; @natcat] Electric fields projected onto specific bonds report on the effect of the surroundings (interacting via coulombic interactions, solvent effects, hydrogen bonding or other forces, all mostly electrostatic in nature) on the flow of electrons along these bonds. Therefore, projected electric fields are correlated to the probability of breaking these bonds, making them a useful probe of chemical reactivity.
+Improved electrostatics with AMOEBA give a unique opportunity to accurately compute electric fields, powerful metrics of catalytic activity in enzymes and other systems [@doi:10.1021/jacs.6b12265; @doi:10.1021/acscatal.7b03151; @natcat]. Electric fields projected onto specific bonds report on the effect of the surroundings (interacting via coulombic interactions, solvent effects, hydrogen bonding or other forces, all mostly electrostatic in nature) on the flow of electrons along these bonds. Therefore, projected electric fields are correlated to the probability of breaking these bonds, making them a useful probe of chemical reactivity.
 
[…]
